safari-600.7-branch/LayoutTests/proximity/create-event-deviceproximity-expected.txt

Tests that document.createEvent() works with deviceproximity

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS deviceproximity event is called properly.
PASS successfullyParsed is true

TEST COMPLETE
